+++ R/map.R
@@ -4,9 +4,11 @@
 #' - `felt_get_map_layers()` returns information about each layer in the map
 #' - `felt_get_map_elements()` returns the shapes for each layer in the map
 #'
-#' @param map_id map identifier from url, from `https://felt.com/map/Readable-Name-map_id`
+#' @param map_id `r template_var_map_id()`
+#' @param layer_id `r template_var_layer_id()`
+#' @param clean `r template_var_clean()`
 #'
-#' @return a [tibble::tibble] for the map
+#' @return a [tibble::tibble] for the map, if `clean = TRUE`, otherwise a list
 #' @export
 #'
 #' @concept get
@@ -18,29 +20,95 @@
 #' # slower, as it has to build the shapes from the API result
 #' felt_get_map_elements('Rockland-2024-Districts-TBI8sDkmQjuK2GX9CSiHiUA')
 #' }
-felt_get_map <- function(map_id) {
+felt_get_map <- function(map_id, clean = TRUE) {
   req <- httr2::request(base_url = api_url()) |>
     httr2::req_url_path_append('maps', map_id) |>
     httr2::req_auth_bearer_token(token = get_felt_key())
 
-  req |>
+  out <- req |>
     httr2::req_perform() |>
-    httr2::resp_body_json() |>
+    httr2::resp_body_json()
+
+  if (!clean) {
+    return(out)
+  }
+
+  out |>
     proc_map()
 }
 
 #' @rdname felt_get_map
 #' @export
-felt_get_map_layers <- function(map_id) {
+felt_get_map_layers <- function(map_id, clean = TRUE) {
   req <- httr2::request(base_url = api_url()) |>
     httr2::req_url_path_append('maps', map_id, 'layers') |>
     httr2::req_auth_bearer_token(token = get_felt_key())
 
-  req |>
+  out <- req |>
     httr2::req_perform() |>
-    httr2::resp_body_json() |>
-    proc_map_layer()
+    httr2::resp_body_json()
+
+  if (!clean) {
+    return(out)
+  }
+
+  out |>
+    list_hoist() |>
+    clean_names()
 }
+
+
+#' @rdname felt_get_map
+#' @export
+felt_get_map_layer <- function(map_id, layer_id, clean = TRUE) {
+  req <- httr2::request(base_url = api_url()) |>
+    httr2::req_url_path_append('maps', map_id, 'layers', layer_id) |>
+    httr2::req_auth_bearer_token(token = get_felt_key())
+
+  out <- req |>
+    httr2::req_perform() |>
+    httr2::resp_body_json()
+
+  if (!clean) {
+    return(out)
+  }
+
+  out |>
+    relist()
+}
+
+#' Get information for a layer group
+#'
+#' @param map_id `r template_var_map_id()`
+#' @param layer_group_id `r template_var_layer_group_id()`
+#' @param clean `r template_var_clean()`
+#'
+#' @return a [tibble::tibble] for the layer group, if `clean = TRUE`, otherwise a list
+#' @export
+#'
+#' @examplesIf has_felt_key()
+#' felt_get_map_layer_group(
+#'   map_id = 'Rockland-2024-Districts-TBI8sDkmQjuK2GX9CSiHiUA',
+#'   layer_group_id = 'rHxyTef7S9CO8W7n1PvBVwC'
+#' )
+felt_get_map_layer_group <- function(map_id, layer_group_id, clean = TRUE) {
+  req <- httr2::request(base_url = api_url()) |>
+    httr2::req_url_path_append('maps', map_id, 'layer_groups', layer_group_id) |>
+    httr2::req_auth_bearer_token(token = get_felt_key())
+
+  out <- req |>
+    httr2::req_perform() |>
+    httr2::resp_body_json()
+
+  if (!clean) {
+    return(out)
+  }
+
+  out |>
+    relist()
+}
+
+
 
 # #' @rdname felt_get_map
 # #' @export