+++ R/user.R
@@ -1,5 +1,7 @@
 #' Obtain information about the user
 #'
+#' @param clean `r template_var_clean()`
+#'
 #' @return a [tibble::tibble] of information about the user
 #' @export
 #'
@@ -7,7 +9,7 @@
 #'
 #' @examplesIf has_felt_key()
 #' felt_get_user()
-felt_get_user <- function() {
+felt_get_user <- function(clean = TRUE) {
   req <- httr2::request(base_url = api_url()) |>
     httr2::req_url_path_append('user') |>
     httr2::req_auth_bearer_token(token = get_felt_key())
@@ -16,10 +18,14 @@ felt_get_user <- function() {
     httr2::req_perform() |>
     httr2::resp_body_json()
 
+  if (!clean) {
+    return(out)
+  }
+
   tibble::tibble(
-    name = out$data$attributes$name,
-    email = out$data$attributes$email,
-    id = out$data$id,
-    type = out$data$type
+    name = out$name,
+    email = out$email,
+    id = out$id,
+    type = out$type
   )
 }